Allow operand types to be reused over instructions
This introduces the Data.Extensible library to provide a sum type of
possible operands. Each instruction takes a sum type of allowed operands
which disallows incorrectly formed instructions at compile time.
A function `doEvalOperand` is provided that will take any such sum time
and evaluate it. So far a 32bit word version of this is implemented